requirements.txt
flask==3.0.0
slack-bolt==1.18.0
slack-sdk==3.23.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
gunicorn==21.2.0
pytest==7.4.3
